Get or Set geotargets Options

Get or set behavior for geospatial data target stores using geotargets-specific global options.

Usage

geotargets_option_get(option_name)

geotargets_option_set(option_name, option_value)

Arguments

option_name

Character. Option name. See Details.

option_value

Value to assign to option x.

Details

Available Options

"geotargets.gdal.raster.driver" - character. Length 1. Set the driver used for raster data in target store (default: "GTiff"). Options for driver names can be found here: https://gdal.org/drivers/raster/index.html

"geotargets.gdal.raster.creation_options" - character. Set the GDAL creation options used when writing raster files to target store (default: ""). You may specify multiple values e.g. c("COMPRESS=DEFLATE", "TFW=YES"). Each GDAL driver supports a unique set of creation options. For example, with the default "GTiff" driver: https://gdal.org/drivers/raster/gtiff.html#creation-options

"geotargets.gdal.vector.driver" - character. Length 1. Set the file type used for vector data in target store (default: "GeoJSON").

"geotargets.gdal.vector.creation_options" - character. Set the GDAL layer creation options used when writing vector files to target store (default: "ENCODING=UTF-8"). You may specify multiple values e.g. c("WRITE_BBOX=YES", "COORDINATE_PRECISION=10"). Each GDAL driver supports a unique set of creation options. For example, with the default "GeoJSON" driver: https://gdal.org/drivers/vector/geojson.html#layer-creation-options

Each option can be overridden with a system environment variable. Options include:

GEOTARGETS_GDAL_RASTER_DRIVER

GEOTARGETS_GDAL_RASTER_CREATION_OPTIONS

GEOTARGETS_GDAL_VECTOR_DRIVER

GEOTARGETS_GDAL_VECTOR_CREATION_OPTIONS

When specifying options that support multiple values using a system environment variable, the separate options should be delimited with a semicolon (";"). For example: "COMPRESS=DEFLATE;TFW=YES".